kill feat: add logic for weather
