+++ src/index.js
@@ -14,3 +14,4 @@ export { renderProfile } from './profile/renderProfilePage.js';
 export { stopTracking } from './map/watchUserPos.js';
 export { renderPrevPage } from './prevPage/renderPrevPage.js';
 export { renderTransportInfo } from './transportInfoPage/renderTransportInfo.js';
+export { displayWeather } from './weather/displayWeather.js';
