--- src/weather/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./style.css">
-    <title>Weather Forecast</title>
-</head>
-
-<body>
-    <div class="weather-card">
-        <div id="weatherContent"></div>
-    </div>
-    <script src="./scripts.js"></script>
-</body>
-
-</html>
