--- src/weather/style.css
@@ -1,75 +1,0 @@
-/*  Поменять классы и подставить переменные  */
-.weather-card {
-    width: 250px;
-    border-radius: 15px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    overflow: hidden;
-    background-color: var(--white);
-    position: relative;
-}
-
-.weather-card-top {
-    background: linear-gradient(to top, #e91e63, #ff9800);
-    color: white;
-    padding: 20px;
-    text-align: left;
-}
-
-.weather-card-top h1 {
-    margin: 0;
-    font-size: 2em;
-    font-weight: normal;
-}
-
-.weather-card-top p {
-    margin: 5px 0;
-    font-size: 1em;
-}
-
-.weather-icon {
-    float: left;
-    margin-right: 10px;
-    font-size: 2em;
-}
-
-.location-icon {
-    float: right;
-    margin-right: 10px;
-    font-size: 1.2em;
-}
-
-.weather-card-bottom {
-    padding: 20px;
-}
-
-.weather-card-bottom table {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-.weather-card-bottom th,
-.weather-card-bottom td {
-    padding: 8px;
-    text-align: left;
-    border-bottom: 1px solid #eee;
-}
-
-.weather-image {
-    width: 100%;
-    height: 76px;
-    display: flex;
-    justify-content: center;
-    background-size: cover;
-    background-position: center;
-}
-
-.weather-description {
-    font-style: italic;
-    margin-bottom: 10px;
-}
-
-.error {
-    color: #d32f2f;
-    padding: 20px;
-    text-align: center;
-}
